requirements.txt
streamlit~=1.27.2
pandas~=2.1.1
requests~=2.31.0
Pillow~=10.0.1
mediapipe~=0.10.5
numpy~=1.26.0
opencv-python~=4.8.1.78
gradio~=3.46.1